integrate dashboardswiper into dashboard and create new scss file
