+++ src/Components/Dashboard/Dashboard.scss
@@ -73,25 +73,6 @@ html {
   margin-bottom: 2vh;
 }
 
-.mySwiper {
-  width: 100%;
-  height: 50vh;
-  z-index: 1;
-  position: relative;
-}
-
-.swiper-slide {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.swiper-slide img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
 .favorites-section {
   @include centered(column, center);
   height: fit-content;
