+++ src/Components/DashboardSwiper/DashboardSwiper.css
@@ -1,0 +1,19 @@
+.mySwiper {
+  width: 100%;
+  height: 50vh;
+  z-index: 1;
+  position: relative;
+}
+
+.swiper-slide {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.swiper-slide img {
+  width: 100%;
+  height: 100%;
+  -o-object-fit: cover;
+     object-fit: cover;
+}/*# sourceMappingURL=DashboardSwiper.css.map */
