+++ src/Components/DashboardSwiper/DashboardSwiper.scss
@@ -1,0 +1,20 @@
+@use "../../../abstract/variables" as *;
+
+.mySwiper {
+  width: 100%;
+  height: 50vh;
+  z-index: 1;
+  position: relative;
+}
+
+.swiper-slide {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.swiper-slide img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
